pyproject.toml
[project]
name = "coursegen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=0.2.6",
    "langchain>=0.3.7",
    "pdfminer-six[image]>=20240706",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
]
